☀️ 20% – Run return-type recovery over the main image too
The rule is not overlay-specific: a callee's declared return type decides the
order of the pre-call argument moves whether the caller is an overlay row or a
main-image owner. Swept over all 755 main-image semantic owners with
`candidate-show` as the oracle, it closes `08093054` -- an owner an earlier
session had recorded as stuck at two halfwords after 74 mode combinations --
and moves 27 others closer, several by tens of halfwords.

`08093054` is the useful witness. It was parked as a scheduling residual and it
was a prototype: one declaration, one flip. Two halfwords of "the allocator
decided" turned out to be an interface we had guessed wrong.

267,272 -> 267,374 bytes. `make verify` green.

+++ semantic/080992f0.c
@@ -30,7 +30,7 @@ struct Object_080992f0 {
     s16 phase_64;
 };
 
-s32 Func_08002322(s32 angle);
+u8 Func_08002322(s32 angle);
 
 void Func_080992f0(struct Object_080992f0 *object)
 {
